refactor: Restructure complete pipeline into packages
