+++ src/main/java/org/molgenis/core/MySpringBootApplication.java
@@ -1,4 +1,4 @@
-package org.molgenis;
+package org.molgenis.core;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
