+++ src/main/java/org/molgenis/validators/ReferenceSequenceValidator.java
@@ -1,4 +1,4 @@
-package org.molgenis;
+package org.molgenis.validators;
 
 import java.util.Map;
 import org.springframework.stereotype.Component;
@@ -18,7 +18,7 @@ public class ReferenceSequenceValidator {
     return true;
   }
 
-  void validateOriginalRef(Map<String, Object> body) {
+  public void validateOriginalRef(Map<String, Object> body) {
     if (!body.containsKey("error") && body.containsKey("ref_orig")) {
       String ref = (String) body.get("ref");
       String refOrig = (String) body.get("ref_orig");
