+++ src/test/java/org/molgenis/mappers/LumcMapperTest.java
@@ -1,4 +1,4 @@
-package org.molgenis;
+package org.molgenis.mappers;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
+import org.molgenis.utils.HgvsService;
 
 class LumcMapperTest {
