+++ src/test/java/org/molgenis/validators/UniquenessCheckerTest.java
@@ -1,4 +1,4 @@
-package org.molgenis;
+package org.molgenis.validators;
 
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
@@ -8,6 +8,7 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import org.molgenis.validators.UniquenessChecker;
 
 class UniquenessCheckerTest {
